Address PR review: strict layer format contract, split flow files, dedupe FileFinder
- calibrate now errors loudly if a calibration job's actual output
  doesn't match the requested layer_data_format, instead of silently
  detecting/adapting to whatever format was produced.
- CalibrationLayer gains is_self_contained_format/
  prepare_metadata_for_output_format so convert.py no longer hand-rolls
  CDF-specific metadata handling.
- FileFinder.find_layers_by_date_and_patterns is now a thin wrapper
  around find_layers_by_patterns instead of a separate implementation.
- convert.py gains debug/info logging throughout.
- performCalibration.py split into one file per flow
  (calibrateFlow/applyFlow/calibrateAndApplyFlow/calibrateConvertFlow)
  plus a calibrationFlowCommon module for shared helpers; tests split
  to match.
- CLAUDE.md: note that tests/unit and tests/integration are separate
  suites and both need checking for changes like a new default format.
