+++ src/prefect_server/workflow.py
@@ -12,16 +12,14 @@ from prefect.schedules import Cron
 from prefect.variables import Variable
 
 from imap_mag.util import CONSTANTS
+from prefect_server.applyFlow import apply_flow
+from prefect_server.calibrateAndApplyFlow import calibrate_and_apply_flow
+from prefect_server.calibrateConvertFlow import calibrate_convert_flow
+from prefect_server.calibrateFlow import calibrate_flow
 from prefect_server.checkIALiRT import check_ialirt_flow
 from prefect_server.constants import PREFECT_CONSTANTS
 from prefect_server.datastoreCleanupFlow import cleanup_datastore_flow
 from prefect_server.datastoreIndexerFlow import index_datastore_flow
-from prefect_server.performCalibration import (
-    apply_flow,
-    calibrate_and_apply_flow,
-    calibrate_convert_flow,
-    calibrate_flow,
-)
 from prefect_server.pollHiEsaStep import (
     poll_hi45_esa_step_flow,
     poll_hi90_esa_step_flow,
