+++ tests/integration/test_performCalibration.py
@@ -7,10 +7,8 @@ from mag_toolkit.calibration.CalibrationConfig import (
     GradiometryConfig,
     ScienceFileVersionConfig,
 )
-from prefect_server.performCalibration import (
-    apply_flow,
-    calibrate_and_apply_flow,
-)
+from prefect_server.applyFlow import apply_flow
+from prefect_server.calibrateAndApplyFlow import calibrate_and_apply_flow
 from tests.util.miscellaneous import open_cdf
 from tests.util.prefect_test_utils import prefect_test_fixture  # noqa: F401
 
@@ -119,7 +117,7 @@ def test_calibrate_and_apply_flow_with_version_overrides(
     )
 
     with patch(
-        "prefect_server.performCalibration._run_calibration",
+        "prefect_server.calibrateAndApplyFlow._run_calibration",
         return_value=[layer_path],
     ):
         calibrate_and_apply_flow.fn(
